animation-iteration-count property is a number instead of integer.

+++ components/style/animation.rs
@@ -30,7 +30,7 @@ use values::computed::Time;
 pub enum KeyframesIterationState {
     Infinite,
     // current, max
-    Finite(u32, u32),
+    Finite(f32, f32),
 }
 
 /// This structure represents wether an animation is actually running.
@@ -90,7 +90,7 @@ impl KeyframesAnimationState {
         }
 
         if let KeyframesIterationState::Finite(ref mut current, ref max) = self.iteration_state {
-            *current += 1;
+            *current += 1.0;
             // NB: This prevent us from updating the direction, which might be
             // needed for the correct handling of animation-fill-mode.
             if *current >= *max {
@@ -460,7 +460,7 @@ pub fn maybe_start_animations(context: &SharedStyleContext,
             let duration = box_style.animation_duration_mod(i).seconds();
             let iteration_state = match box_style.animation_iteration_count_mod(i) {
                 AnimationIterationCount::Infinite => KeyframesIterationState::Infinite,
-                AnimationIterationCount::Number(n) => KeyframesIterationState::Finite(0, n),
+                AnimationIterationCount::Number(n) => KeyframesIterationState::Finite(0.0, n),
             };
 
             let animation_direction = box_style.animation_direction_mod(i);
